Removing KernelAgent.cs and moving its functionality into Agent.cs
