+++ dotnet/src/Agents/AzureAI/AzureAIAgent.cs
@@ -16,9 +16,9 @@ using AAIP = Azure.AI.Projects;
 namespace Microsoft.SemanticKernel.Agents.AzureAI;
 
 /// <summary>
-/// Provides a specialized <see cref="KernelAgent"/> based on an Azure AI agent.
+/// Provides a specialized <see cref="Agent"/> based on an Azure AI agent.
 /// </summary>
-public sealed partial class AzureAIAgent : KernelAgent
+public sealed partial class AzureAIAgent : Agent
 {
     /// <summary>
     /// Provides tool definitions used when associating a file attachment to an input message:
