+++ dotnet/src/Agents/Bedrock/BedrockAgent.cs
@@ -16,9 +16,9 @@ using Microsoft.SemanticKernel.Diagnostics;
 namespace Microsoft.SemanticKernel.Agents.Bedrock;
 
 /// <summary>
-/// Provides a specialized <see cref="KernelAgent"/> for the Bedrock Agent service.
+/// Provides a specialized <see cref="Agent"/> for the Bedrock Agent service.
 /// </summary>
-public sealed class BedrockAgent : KernelAgent
+public sealed class BedrockAgent : Agent
 {
     /// <summary>
     /// The client used to interact with the Bedrock Agent service.
