+++ dotnet/src/Functions/Functions.OpenApi.Extensions/Extensions/DeclarativeAgentExtensions.cs
@@ -35,7 +35,7 @@ public static class DeclarativeAgentExtensions
         CopilotAgentPluginParameters? pluginParameters = null,
         PromptExecutionSettings? promptExecutionSettings = default,
         CancellationToken cancellationToken = default)
-        where T : KernelAgent, new()
+        where T : Agent, new()
     {
         Verify.NotNull(kernel);
         Verify.NotNullOrWhiteSpace(filePath);
